Add support for BasicObject
Apply instance_variable_set to Object and instance_eval to BasicObject

Add test case for non BasicObject/Object classes

Move BasicObject and Object related logic into separate modules [re-run specs]
