+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "support/coverage"
+require_relative "support/shared_examples/memoizable"
 
 $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
 require "rspec/version"
